refactor: improve backend routers, schemas, services, and utils
- Update 70+ routers for permission guards, error handling, and pagination
- Enhance schemas with Decimal precision for monetary fields
- Fix services: GL, tax engine, inventory WAC, payroll, recurring JE
- Improve utils: masking, permissions, treasury balance, ZATCA clearance
- Add i18n error messages for new validation rules

+++ backend/routers/inventory/schemas.py
@@ -82,15 +82,15 @@ class SupplierCreate(BaseModel):
 
 class SupplierResponse(SupplierCreate):
     id: int
-    current_balance: float
-    balance: Optional[float] = None
-    balance_bc: Optional[float] = None
+    current_balance: Decimal
+    balance: Optional[Decimal] = None
+    balance_bc: Optional[Decimal] = None
     is_active: bool
     currency: Optional[str] = None
     created_at: datetime
     display_currency: Optional[str] = None
-    balance_display: Optional[float] = None
-    balance_sar: Optional[float] = None
+    balance_display: Optional[Decimal] = None
+    balance_sar: Optional[Decimal] = None
     balances: Optional[list] = None
     sites: Optional[list] = None
     site_id: Optional[int] = None
