+++ backend/schemas/blanket_po.py
@@ -2,12 +2,13 @@
 from pydantic import BaseModel, Field
 from typing import Optional
 from datetime import date
+from decimal import Decimal
 
 
 class BlanketPOCreate(BaseModel):
     supplier_id: int
-    total_quantity: float = Field(gt=0)
-    unit_price: float = Field(gt=0)
+    total_quantity: Decimal = Field(gt=0)
+    unit_price: Decimal = Field(gt=0)
     valid_from: date
     valid_to: date
     branch_id: Optional[int] = None
@@ -20,13 +21,13 @@ class BlanketPORead(BaseModel):
     id: int
     supplier_id: int
     agreement_number: str
-    total_quantity: float
-    unit_price: float
-    total_amount: float
-    released_quantity: float
-    released_amount: float
-    remaining_quantity: float
-    remaining_amount: float
+    total_quantity: Decimal
+    unit_price: Decimal
+    total_amount: Decimal
+    released_quantity: Decimal
+    released_amount: Decimal
+    remaining_quantity: Decimal
+    remaining_amount: Decimal
     valid_from: Optional[date] = None
     valid_to: Optional[date] = None
     status: str
@@ -42,12 +43,12 @@ class BlanketPORead(BaseModel):
 
 
 class ReleaseOrderCreate(BaseModel):
-    release_quantity: float = Field(gt=0)
+    release_quantity: Decimal = Field(gt=0)
     release_date: Optional[date] = None
     notes: Optional[str] = None
 
 
 class PriceAmendRequest(BaseModel):
-    new_price: float = Field(gt=0)
+    new_price: Decimal = Field(gt=0)
     effective_date: date
     reason: Optional[str] = None
